Don't delete inactive sessions in MFA process, allow to cancel MFA process #795 #802 #801

+++ src/main/java/app/nzyme/core/security/authentication/AuthenticationService.java
@@ -755,16 +755,21 @@ public class AuthenticationService {
 
             // Delete all sessions of users that have been inactive for 15 minutes.
             List<Long> inactiveUsers = nzyme.getDatabase().withHandle(handle ->
-                    handle.createQuery("SELECT id FROM auth_users WHERE last_activity < :timeout")
+                    handle.createQuery("SELECT u.id FROM auth_users AS u " +
+                                    "LEFT JOIN auth_sessions AS s ON s.user_id = u.id " +
+                                    "WHERE s.mfa_valid = true AND u.last_activity < :timeout")
                             .bind("timeout", DateTime.now().minusMinutes(15))
                             .mapTo(Long.class)
                             .list()
             );
-            nzyme.getDatabase().useHandle(handle ->
-                    handle.createUpdate("DELETE FROM auth_sessions WHERE user_id IN (<user_ids>)")
-                            .bindList("user_ids", inactiveUsers)
-                            .execute()
-            );
+
+            if (!inactiveUsers.isEmpty()) {
+                nzyme.getDatabase().useHandle(handle ->
+                        handle.createUpdate("DELETE FROM auth_sessions WHERE user_id IN (<user_ids>)")
+                                .bindList("user_ids", inactiveUsers)
+                                .execute()
+                );
+            }
         } catch(Exception e) {
             LOG.error("Could not run session cleaning.", e);
         }
